server/requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask-CORS==5.0.0
gunicorn